+++ exe/atlaspeeringctl
@@ -1,10 +1,10 @@
 #!/usr/bin/env ruby
 # frozen_string_literal: true
 
-gem_lib = File.expand_path('../../lib', __FILE__)
-$LOAD_PATH.unshift(gem_lib) unless $LOAD_PATH.include?(gem_lib)
+lib = File.expand_path('../../lib', __FILE__)
+$LOAD_PATH.unshift(lib) unless $LOAD_PATH.include?(lib)
 
 require 'atlas'
 
 Process.setproctitle('atlas_peering_ctl')
-Atlas::Peering::CLI.run(ARGV)
+Atlas::CLI.run(ARGV)
