--- lib/atlas/peering/commander.rb
@@ -1,31 +1,0 @@
-# frozen_string_literal: true
-
-module Atlas
-  module Peering
-    ##
-    # class: Atlas::Peering::CommandNotImplemented - The convenient exception
-    class CommandNotImplemented < StandardError
-    end
-
-    ##
-    # class: Atlas::Peering::Commander - The command factory class
-    class Commander
-      def initialize(client)
-        @client = client
-
-        self
-      end
-
-      def launch(command, data)
-        case command
-        when :list then
-          @client.list(data)
-        when :create then
-          @client.create(data)
-        else
-          raise CommandNotImplemented.new(command)
-        end
-      end
-    end
-  end
-end
