--- lib/atlas/peering/logger.rb
@@ -1,23 +1,0 @@
-# frozen_string_literal: true
-
-module Atlas
-  module Peering
-    ##
-    # class: Atlas::Peering::Logger: The application logger
-    class Logger
-      LOG_LEVELS = {
-        debug: ::Logger::DEBUG,
-        info:  ::Logger::INFO,
-        warn:  ::Logger::WARN,
-        error: ::Logger::ERROR
-      }.freeze
-
-      def self.configure(level, target = STDOUT)
-        logger       = ::Logger.new(target)
-        logger.level = LOG_LEVELS.fetch(level, ::Logger::INFO)
-
-        logger
-      end
-    end
-  end
-end
