+++ lib/atlas/version.rb
@@ -1,7 +1,5 @@
 # frozen_string_literal: true
 
 module Atlas
-  module Peering
-    VERSION = '0.2.0'
-  end
+  VERSION = '0.2.0'
 end
